updated tests to task-sdk/tests directory
